+++ app/src/main/java/com/mkumar/data/ProductFormData.kt
@@ -10,7 +10,7 @@ import kotlinx.serialization.json.JsonClassDiscriminator
 @JsonClassDiscriminator("type") 
 sealed class ProductFormData {
 
-    abstract val unitPrice: Long
+    abstract val unitPrice: Int
     abstract val quantity: Int
     abstract val discountPct: Int
     abstract val total: Int
@@ -21,7 +21,7 @@ sealed class ProductFormData {
         val brand: String = "",
         val color: String = "",
         val size: String = "",
-        override val unitPrice: Long = 0L,
+        override val unitPrice: Int = 0,
         override val quantity: Int = 1,
         override val discountPct: Int = 0,
         override val total: Int = 0
@@ -34,7 +34,7 @@ sealed class ProductFormData {
         val leftAxis: String = "",
         val rightSphere: String = "",
         val rightAxis: String = "",
-        override val unitPrice: Long = 0L,
+        override val unitPrice: Int = 0,
         override val quantity: Int = 1,
         override val discountPct: Int = 0,
         override val total: Int = 0
@@ -45,7 +45,7 @@ sealed class ProductFormData {
     data class ContactLensData(
         val power: String = "",
         val duration: String = "",
-        override val unitPrice: Long = 0L,
+        override val unitPrice: Int = 0,
         override val quantity: Int = 1,
         override val discountPct: Int = 0,
         override val total: Int = 0
